requirements.txt
torch
torchaudio
transformers
librosa
numpy
pandas
pandarallel
tqdm
scikit-learn
jiwer
pyyaml